Rename part 1.asm to part1.asm
Hangman is a simple 2-player game where one player picks a word or phrase and the other player tries to guess it. The word to guess is marked out by underscores, where each underscore represents one letter. The guesser picks one letter at a time. If the letter is in the word then all occurrences of that letter replace the dash(es) at the appropriate place(s). The game continues until either the guesser picks all the letters in the word, or until they reach a predetermined limit of incorrect guesses.
(a) Write an assembly program that reads a word entered by the player. The word is composed of one or more lowercase letters only followed by a new line character. Then read a single, lowercase letter followed by a newline character which is a guess. Print the state of the game after one guess, where if the guessed letter appears in the word then it is printed in its correct spot(s), otherwise a underscore is printed in place of the letter in the word. A newline character should be printed at the end of the state. The ASCII code value of a newline character is the hex number 0xA.
For example, if the user enters the word waterloo, then enters the letter o, the following characters will be printed on their own line:
_ _ _ _ _ _oo
Test your solution using the noargs front end.